Add files via upload
This Power BI project revolves around assembling a cricket team consisting of 11 players from various parts of the world. The selection process is based on the players' performances in the cricket T20 World Cup. The team formation involves picking 4 top batsmen, 4 top bowlers, 1 wicketkeeper, and a few all-rounders.

To initiate the project, data is gathered from the T20 World Cup's official website, and this is achieved using the Bright Data tool, which extracts data in JSON file format. Once collected, the data is imported into Python, utilizing libraries such as Pandas and JSON for data manipulation and transformation. After the necessary cleaning and structuring of the data, it is converted into a CSV file format.

The subsequent stages of the project occur within Power BI. This involves importing the preprocessed data in CSV format into Power BI for further analysis and visualization. In Power BI, the data can be transformed, modeled, and visualized to provide insights into player performance, team composition, and other relevant factors. This could include creating graphs, charts, and tables to showcase the statistics and characteristics of the selected players.

In essence, this Power BI project leverages data from the T20 World Cup, first gathering it using the Bright Data tool and then refining it in Python before moving onto Power BI for the final stages of analysis and visualization to construct the optimal cricket team based on the players' T20 World Cup performances.